Flying special workers spawn workaround.
Workaround for allowing to spawn special workers with fly ability.

+++ src/thing_list.c
@@ -1675,13 +1675,13 @@ TbBool update_thing(struct Thing *thing)
     {
         if ((thing->field_1 & 0x04) != 0)
         {
-          thing->pos_2C.x.val += thing->acceleration.x.val;
-          thing->pos_2C.y.val += thing->acceleration.y.val;
-          thing->pos_2C.z.val += thing->acceleration.z.val;
-          thing->acceleration.x.val = 0;
-          thing->acceleration.y.val = 0;
-          thing->acceleration.z.val = 0;
-          set_flag_byte(&thing->field_1, 0x04, false);
+            thing->pos_2C.x.val += thing->acceleration.x.val;
+            thing->pos_2C.y.val += thing->acceleration.y.val;
+            thing->pos_2C.z.val += thing->acceleration.z.val;
+            thing->acceleration.x.val = 0;
+            thing->acceleration.y.val = 0;
+            thing->acceleration.z.val = 0;
+            set_flag_byte(&thing->field_1, 0x04, false);
         }
         thing->velocity.x.val = thing->pos_2C.x.val;
         thing->velocity.y.val = thing->pos_2C.y.val;
@@ -1711,26 +1711,26 @@ TbBool update_thing(struct Thing *thing)
     {
         if (thing->mappos.z.val > thing->field_60)
         {
-          if (thing->pos_2C.x.val != 0)
-            thing->pos_2C.x.val = thing->pos_2C.x.val * (long)(256 - thing->field_24) / 256;
-          if (thing->pos_2C.y.val != 0)
-            thing->pos_2C.y.val = thing->pos_2C.y.val * (long)(256 - thing->field_24) / 256;
-          if ((thing->movement_flags & TMvF_Flying) == 0)
-          {
-            thing->acceleration.z.val -= thing->field_20;
-            thing->field_1 |= 0x04;
-          }
+            if (thing->pos_2C.x.val != 0)
+              thing->pos_2C.x.val = thing->pos_2C.x.val * (256 - (int)thing->field_24) / 256;
+            if (thing->pos_2C.y.val != 0)
+              thing->pos_2C.y.val = thing->pos_2C.y.val * (256 - (int)thing->field_24) / 256;
+            if ((thing->movement_flags & TMvF_Flying) == 0)
+            {
+                thing->acceleration.z.val -= thing->field_20;
+                thing->field_1 |= 0x04;
+            }
         } else
         {
-          if (thing->pos_2C.x.val != 0)
-            thing->pos_2C.x.val = thing->pos_2C.x.val * (long)(256 - thing->field_23) / 256;
-          if (thing->pos_2C.y.val != 0)
-            thing->pos_2C.y.val = thing->pos_2C.y.val * (long)(256 - thing->field_23) / 256;
-          thing->mappos.z.val = thing->field_60;
-          if ((thing->movement_flags & TMvF_Unknown08) != 0)
-          {
-            thing->pos_2C.z.val = 0;
-          }
+            if (thing->pos_2C.x.val != 0)
+              thing->pos_2C.x.val = thing->pos_2C.x.val * (256 - (int)thing->field_23) / 256;
+            if (thing->pos_2C.y.val != 0)
+              thing->pos_2C.y.val = thing->pos_2C.y.val * (256 - (int)thing->field_23) / 256;
+            thing->mappos.z.val = thing->field_60;
+            if ((thing->movement_flags & TMvF_Unknown08) != 0)
+            {
+              thing->pos_2C.z.val = 0;
+            }
         }
     }
     update_thing_animation(thing);
